requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
requests==2.31.0
python-multipart==0.0.6
sentence-transformers==2.2.2
scikit-learn==1.3.2
azure-storage-file-datalake==12.12.0
azure-storage-blob==12.18.0
azure-identity==1.14.0
azure-mgmt-cognitiveservices==13.5.0
openai==1.10.0
numpy==1.26.0
httpx==0.24.1
